refactoring - moved code from views to separate files
