+++ courses/urls.py
@@ -1,48 +1,50 @@
 from django.urls import path
-from . import views
+
+from .views import course
+from .views import homework
+from .views import project
 
 
 urlpatterns = [
-    # path('ping', views.ping, name='ping'),
-    path("", views.course_list, name="course_list"),
+    path("", course.course_list, name="course_list"),
     path(
         "<slug:course_slug>/",
-        views.course_view,
+        course.course_view,
         name="course",
     ),
     path(
         "<slug:course_slug>/leaderboard",
-        views.leaderboard_view,
+        course.leaderboard_view,
         name="leaderboard",
     ),
     path(
         "<slug:course_slug>/leaderboard/<int:enrollment_id>/",
-        views.leaderboard_score_breakdown_view,
+        course.leaderboard_score_breakdown_view,
         name="leaderboard_score_breakdown",
     ),
     path(
         "<slug:course_slug>/enrollment",
-        views.enrollment_view,
+        course.enrollment_view,
         name="enrollment",
     ),
     path(
         "<slug:course_slug>/project/<slug:project_slug>",
-        views.project_view,
+        project.project_view,
         name="project",
     ),
     path(
         "<slug:course_slug>/project/<slug:project_slug>/eval",
-        views.projects_eval_view,
+        project.projects_eval_view,
         name="projects_eval",
     ),
     path(
         "<slug:course_slug>/project/<slug:project_slug>/eval/<int:review_id>",
-        views.projects_eval_submit,
+        project.projects_eval_submit,
         name="projects_eval_submit",
     ),
     path(
         "<slug:course_slug>/homework/<slug:homework_slug>",
-        views.homework_view,
+        homework.homework_view,
         name="homework",
     ),
 ]
